Add files via upload
This is a case designed to add the 1lc Duet Toolboard V1.1 to the LGX extruder for the RatRig V-Core extruder system. Printing the lid is pretty straightforward... the bottom part is a little tricky. I print in the same orientation as it would be on the printer and then use supports. Also before anyone comments that the Board doesn't look centered on the Extruder, this is because the Board itself is designed to mount to the Hemera from E3D. So the bolt pattern on the board is off center so it looks a little wonky. But My goal here was to get the board between the fans and the motor. It's a tight fit big time... So that's why the lid is so thin. This originally was based on the existing design by Maciej Raczyński. However I just scratched his design and made my own. But it is similar in spirit.